requirements.txt
base58==2.1.1
ecdsa==0.17.0
pycryptodome==3.14.1
shellescape==3.8.1
six==1.16.0
